remove : 테스트 주석 및 dto 변환 삭제, 테스트 코드 이동
