+++ src/main/java/com/poppin/backendserver/controller/LocationController.java
@@ -31,7 +31,6 @@ public class LocationController {
 
     /**
      * 검색 및 페이징
-     * 테스트완료
      */
     @GetMapping("/search")
         public Page<LocationDTO> searchLocation(@RequestParam("keyword") String keyword, @RequestParam("page") int page, @RequestParam("size") int size) {
@@ -42,13 +41,12 @@ public class LocationController {
 
     /**
      * 특정 가게 정보 가져오기
-     * 테스트완료
      */
     @GetMapping("/{location_id}")
     public ResponseEntity<LocationDTO> getLocation(@PathVariable("location_id") Long id) {
         Optional<Location> location = locationService.getLocation(id);
         if(location.isPresent()) {
-            LocationDTO locationDTO = convertToDTO(location.get());
+            LocationDTO locationDTO = new LocationDTO(location.get());
             return ResponseEntity.ok(locationDTO);
         }
         return ResponseEntity.notFound().build();
@@ -56,7 +54,6 @@ public class LocationController {
 
     /**
      * 가게의 리뷰 가져오기
-     * 테스트 완료
      */
     @GetMapping("/{location_id}/reviews")
     public ResponseEntity<List<ReviewDTO>> getReviews(@PathVariable("location_id") Long id) {
@@ -68,72 +65,10 @@ public class LocationController {
 
     /**
      * 가게 생성
-     * 테스트완료
      */
     @PostMapping("/create")
     public ResponseEntity<Long> createLocation(@RequestBody LocationDTO locationDTO) {
-        Location location = convertToEntity(locationDTO);
-        Long savedLocationId = locationService.saveLocation(location);
-        return ResponseEntity.ok(savedLocationId);
-    }
-
-    /**
-     * 가게 정보 업데이트
-     * 테스트완료
-     */
-    @PutMapping("/{location_id}")
-    public ResponseEntity<Long> updateLocation(@PathVariable("location_id") Long id, @RequestBody LocationDTO locationDTO) {
-        locationService.updateLocation(id, convertToEntity(locationDTO));
-        return ResponseEntity.ok(id);
-    }
-
-    /**
-     * 가게 삭제
-     * 테스트완료
-     */
-    @DeleteMapping("/{location_id}")
-    public ResponseEntity<Long> deleteLocation(@PathVariable("location_id") Long id) {
-        Long deletedLocationId = locationService.deleteLocation(id);
-        return ResponseEntity.ok(deletedLocationId);
-    }
-
-    /**
-     * 테스트 위한 더미데이터
-     */
-    @PostConstruct
-    public void init(){
-        for (int i = 0; i < 100; i++) {
-            locationService.saveLocation(Location.builder()
-                    .name("apple"+i)
-                    .address("korea"+i)
-                    .telephone("0101234"+i)
-                    .closedDay("monday")
-                    .openTime("10~21")
-                    .latitude((long) (10+i))
-                    .lontitude((long) (30+i))
-                    .build());
-        }
-    }
-
-    private LocationDTO convertToDTO(Location location) {
-        LocationDTO locationDTO = LocationDTO.builder()
-                .id(location.getId())
-                .name(location.getName())
-                .address(location.getAddress())
-                .telephone(location.getTelephone())
-                .ratingSum(location.getRatingSum())
-                .ratingCount(location.getRatingCount())
-                .openTime(location.getOpenTime())
-                .closedDay(location.getClosedDay())
-                .latitude(location.getLatitude())
-                .lontitude(location.getLontitude())
-                .build();
-        return locationDTO;
-    }
-
-    private Location convertToEntity(LocationDTO locationDTO) {
         Location location = Location.builder()
-                .id(locationDTO.getId())
                 .name(locationDTO.getName())
                 .address(locationDTO.getAddress())
                 .telephone(locationDTO.getTelephone())
@@ -144,6 +79,31 @@ public class LocationController {
                 .latitude(locationDTO.getLatitude())
                 .lontitude(locationDTO.getLontitude())
                 .build();
-        return location;
+        Long savedLocationId = locationService.saveLocation(location);
+        return ResponseEntity.ok(savedLocationId);
     }
+
+    /**
+     * 가게 정보 업데이트
+     */
+    @PutMapping("/{location_id}")
+    public ResponseEntity<Long> updateLocation(@PathVariable("location_id") Long id, @RequestBody LocationDTO locationDTO) {
+        Location location = Location.builder()
+                .telephone(locationDTO.getTelephone())
+                .closedDay(locationDTO.getClosedDay())
+                .openTime(locationDTO.getOpenTime())
+                .build();
+        locationService.updateLocation(id, location);
+        return ResponseEntity.ok(id);
+    }
+
+    /**
+     * 가게 삭제
+     */
+    @DeleteMapping("/{location_id}")
+    public ResponseEntity<Long> deleteLocation(@PathVariable("location_id") Long id) {
+        Long deletedLocationId = locationService.deleteLocation(id);
+        return ResponseEntity.ok(deletedLocationId);
+    }
+
 }
