+++ src/main/java/com/poppin/backendserver/controller/ReviewController.java
@@ -24,13 +24,12 @@ public class ReviewController {
 
     /**
      * 특정 리뷰 가져오기
-     * 테스트완료
      */
     @GetMapping("/{review_id}")
     public ResponseEntity<ReviewDTO> getReview(@PathVariable("review_id") Long id){
         Optional<Review> review = reviewService.getReview(id);
         if (review.isPresent()) {
-            ReviewDTO reviewDTO = convertToDTO(review.get());
+            ReviewDTO reviewDTO = new ReviewDTO(review.get());
             return ResponseEntity.ok(reviewDTO);
         }
         return ResponseEntity.notFound().build();
@@ -38,27 +37,35 @@ public class ReviewController {
 
     /**
      * 리뷰 생성
-     * 테스트완료
      */
     @PostMapping("/create")
     public ResponseEntity<Long> createReview(@RequestBody ReviewDTO reviewDTO) {
-        Long savedReviewId = reviewService.saveReview(convertToEntity(reviewDTO));
+        Review review = Review.builder()
+                .rating(reviewDTO.getRating())
+                .title(reviewDTO.getTitle())
+                .text(reviewDTO.getText())
+                .location(locationService.getLocation(reviewDTO.getLocationId()).get())
+                .build();
+        Long savedReviewId = reviewService.saveReview(review);
         return ResponseEntity.ok(savedReviewId);
     }
 
     /**
      * 특정 리뷰 업데이트
-     * 테스트완료
      */
     @PutMapping("/{review_id}")
     public ResponseEntity<Long> updateReview(@PathVariable("review_id") Long id, @RequestBody ReviewDTO reviewDTO) {
-        reviewService.updateReview(id, convertToEntity(reviewDTO));
+        Review review = Review.builder()
+                .rating(reviewDTO.getRating())
+                .title(reviewDTO.getTitle())
+                .text(reviewDTO.getText())
+                .build();
+        reviewService.updateReview(id, review);
         return ResponseEntity.ok(id);
     }
 
     /**
      * 특정 리뷰 삭제
-     * 테스트완료
      */
     @DeleteMapping("/{review_id}")
     public ResponseEntity<Long> deleteReview(@PathVariable("review_id") Long id) {
@@ -66,49 +73,4 @@ public class ReviewController {
         return ResponseEntity.ok(deletedReviewId);
     }
 
-    /**
-     * 테스트 위한 더미데이터
-     */
-    @PostConstruct
-    public void init(){
-        Location location = Location.builder()
-                .name("appler")
-                .address("korear")
-                .telephone("0101234r")
-                .closedDay("monday")
-                .openTime("10~21")
-                .latitude((long) (101))
-                .lontitude((long) (303))
-                .build();
-        locationService.saveLocation(location);
-        for (int i = 0; i < 100; i++) {
-            reviewService.saveReview(Review.builder()
-                    .rating(4)
-                    .title("Nice food"+i)
-                    .text("so delicius"+i)
-                    .location(location)
-                    .build());
-        }
-    }
-    private ReviewDTO convertToDTO(Review review) {
-        ReviewDTO reviewDTO = ReviewDTO.builder()
-                .id(review.getId())
-                .rating(review.getRating())
-                .title(review.getTitle())
-                .text(review.getText())
-                .locationId(review.getLocation().getId())
-//                .userId(review.getUser().getId())
-                .build();
-        return reviewDTO;
-    }
-    private Review convertToEntity(ReviewDTO reviewDTO) {
-        Review review = Review.builder()
-                .id(reviewDTO.getId())
-                .rating(reviewDTO.getRating())
-                .title(reviewDTO.getTitle())
-                .text(reviewDTO.getText())
-                .location(locationService.getLocation(reviewDTO.getLocationId()).get())
-                .build();
-        return review;
-    }
 }
